Refactor fallback module.
actually not working in ie yet but i want to merge to master
anyway on the off chance that you guys (hi Brendan, Paul, and
Justin) look at this.  i have to work an amination [sic] job
tomorrow so i'm concerned i won't have a chance to address
whatever is going on with ie8 until fri AM at earliest.
i have to get a proper VM going too... that will help.
browserstack is not quite cutting it.